+++ Assets/PatchKit Patcher/Scripts/Status/GeneralStatusHolder.cs
@@ -5,6 +5,8 @@
         public double Weight { get; private set; }
 
         public double Progress { get; set; }
+        
+        public string Description { get; set; }
 
         public GeneralStatusHolder(double weight)
         {